Update detail page on edit Fix DB to use long for id, not int Make recyclerview adapter static for use in add/edit Dialog Adding and editing Restaurants now changes RestaurantListContent and DB Improve RecyclerView performance Make RecyclerView a Grid

+++ app/src/main/java/nl/verhoogenvansetten/restaurantrio/model/Restaurant.java
@@ -42,7 +42,7 @@ public class Restaurant{
     }
 
     //Constructor which is used to load existing restaurant objects from the DB.
-    public Restaurant(int id, String name, String location, String description, Bitmap image) {
+    public Restaurant(long id, String name, String location, String description, Bitmap image) {
         this.id = id;
         this.name = name;
         this.location = location;
